refactor(approvals): move approval handlers to config module
`add_approvals` and `clear_approvals` no longer back any `wt hook`
subcommand — they're invoked through `wt config approvals`. Move them
into `src/commands/config/approvals.rs` so the file location matches
the user-facing path.

Tidy the related test names: `test_hook_clear_with_approvals` exercised
the legacy `[projects.X]` storage in `config.toml` — move it to
`approvals.rs` and rename to `test_clear_approvals_legacy_config_storage`.
Drop `test_hook_clear_no_approvals` as a duplicate (same command, same
output as `test_clear_approvals_no_approvals`).
