+++ src/commands/config/mod.rs
@@ -2,6 +2,7 @@
 //!
 //! Commands for managing user config, project config, state, and hints.
 
+mod approvals;
 mod create;
 mod hints;
 pub mod opencode;
@@ -11,6 +12,7 @@ mod state;
 mod update;
 
 // Re-export public functions
+pub use approvals::{add_approvals, clear_approvals};
 pub use create::handle_config_create;
 pub use hints::{handle_hints_clear, handle_hints_get};
 pub use opencode::{handle_opencode_install, handle_opencode_uninstall};
