+++ src/commands/hook_commands.rs
@@ -2,8 +2,6 @@
 //!
 //! This module contains:
 //! - `run_hook` - Execute a specific hook type
-//! - `add_approvals` - Approve all project commands
-//! - `clear_approvals` - Clear approved commands
 //! - `handle_hook_show` - Display configured hooks
 
 use std::collections::HashMap;
@@ -14,11 +12,11 @@ use color_print::cformat;
 use strum::IntoEnumIterator;
 use worktrunk::HookType;
 use worktrunk::config::{Approvals, CommandConfig, ProjectConfig, UserConfig};
-use worktrunk::git::{GitError, Repository};
+use worktrunk::git::Repository;
 use worktrunk::path::format_path_for_display;
 use worktrunk::styling::{
     INFO_SYMBOL, PROMPT_SYMBOL, eprintln, format_bash_with_gutter, format_heading, hint_message,
-    info_message, success_message, warning_message,
+    info_message, warning_message,
 };
 
 use super::command_approval::approve_hooks_filtered;
@@ -32,7 +30,6 @@ use super::hooks::{
     HookCommandSpec, check_name_filter_matched, count_sourced_commands, prepare_background_hooks,
     prepare_sourced_steps, run_hook_with_filter, spawn_hook_pipeline,
 };
-use super::project_config::{collect_commands_for_aliases, collect_commands_for_hooks};
 
 fn run_filtered_hook(
     ctx: &CommandContext,
@@ -386,127 +383,6 @@ pub fn run_hook(
             name_filters,
         ),
     }
-}
-
-/// Handle `wt config approvals add` command - approve all hook and alias commands in the project
-pub fn add_approvals(show_all: bool) -> anyhow::Result<()> {
-    use super::command_approval::approve_command_batch;
-
-    let repo = Repository::current()?;
-    let project_id = repo.project_identifier()?;
-    let approvals = Approvals::load().context("Failed to load approvals")?;
-
-    // Load project config (error if missing - this command requires it)
-    let config_path = repo
-        .project_config_path()?
-        .context("Cannot determine project config location — no worktree found")?;
-    let project_config = repo
-        .load_project_config()?
-        .ok_or(GitError::ProjectConfigNotFound { config_path })?;
-
-    // Collect all commands from the project config: hooks first (lifecycle order),
-    // then aliases (alphabetical via BTreeMap).
-    let all_hooks: Vec<_> = HookType::iter().collect();
-    let mut commands = collect_commands_for_hooks(&project_config, &all_hooks);
-    commands.extend(collect_commands_for_aliases(&project_config));
-
-    if commands.is_empty() {
-        eprintln!("{}", info_message("No commands configured in project"));
-        return Ok(());
-    }
-
-    // Filter to only unapproved commands (unless --all is specified)
-    let commands_to_approve = if !show_all {
-        let unapproved: Vec<_> = commands
-            .into_iter()
-            .filter(|cmd| !approvals.is_command_approved(&project_id, &cmd.command.template))
-            .collect();
-
-        if unapproved.is_empty() {
-            eprintln!("{}", info_message("All commands already approved"));
-            return Ok(());
-        }
-
-        unapproved
-    } else {
-        commands
-    };
-
-    // Call the approval prompt (yes=false to require interactive approval and save)
-    // When show_all=true, we've already included all commands in commands_to_approve
-    // When show_all=false, we've already filtered to unapproved commands
-    // So we pass skip_approval_filter=true to prevent double-filtering
-    let approved =
-        approve_command_batch(&commands_to_approve, &project_id, &approvals, false, true)?;
-
-    // Show result
-    if approved {
-        eprintln!("{}", success_message("Commands approved & saved to config"));
-    } else {
-        eprintln!("{}", info_message("Commands declined"));
-    }
-
-    Ok(())
-}
-
-/// Handle `wt config approvals clear` command - clear approved commands
-pub fn clear_approvals(global: bool) -> anyhow::Result<()> {
-    let mut approvals = Approvals::load().context("Failed to load approvals")?;
-
-    if global {
-        // Count projects with approvals before clearing
-        let project_count = approvals
-            .projects()
-            .filter(|(_, cmds)| !cmds.is_empty())
-            .count();
-
-        if project_count == 0 {
-            eprintln!("{}", info_message("No approvals to clear"));
-            return Ok(());
-        }
-
-        approvals
-            .clear_all(None)
-            .context("Failed to clear approvals")?;
-
-        eprintln!(
-            "{}",
-            success_message(format!(
-                "Cleared approvals for {project_count} project{}",
-                if project_count == 1 { "" } else { "s" }
-            ))
-        );
-    } else {
-        // Clear approvals for current project (default)
-        let repo = Repository::current()?;
-        let project_id = repo.project_identifier()?;
-
-        // Count approvals before clearing
-        let approval_count = approvals
-            .projects()
-            .find(|(id, _)| *id == project_id)
-            .map(|(_, cmds)| cmds.len())
-            .unwrap_or(0);
-
-        if approval_count == 0 {
-            eprintln!("{}", info_message("No approvals to clear for this project"));
-            return Ok(());
-        }
-
-        approvals
-            .revoke_project(&project_id, None)
-            .context("Failed to clear project approvals")?;
-
-        eprintln!(
-            "{}",
-            success_message(format!(
-                "Cleared {approval_count} approval{} for this project",
-                if approval_count == 1 { "" } else { "s" }
-            ))
-        );
-    }
-
-    Ok(())
 }
 
 /// Handle `wt hook show` command - display configured hooks
