+++ tests/integration_tests/hook_show.rs
@@ -313,73 +313,6 @@ lint = "pre-commit run"
     });
 }
 
-/// Test `wt config approvals clear` when no approvals exist for the project.
-#[rstest]
-fn test_hook_clear_no_approvals(repo: TestRepo, temp_home: TempDir) {
-    // Remove origin so project_identifier uses full canonical path
-    repo.run_git(&["remote", "remove", "origin"]);
-
-    // Create user config without any project approvals
-    let global_config_dir = temp_home.path().join(".config").join("worktrunk");
-    fs::create_dir_all(&global_config_dir).unwrap();
-    let config_path = global_config_dir.join("config.toml");
-    fs::write(
-        &config_path,
-        r#"worktree-path = "../{{ repo }}.{{ branch }}"
-"#,
-    )
-    .unwrap();
-
-    let settings = setup_snapshot_settings_with_home(&repo, &temp_home);
-    settings.bind(|| {
-        let mut cmd = wt_command();
-        repo.configure_wt_cmd(&mut cmd);
-        cmd.env("WORKTRUNK_CONFIG_PATH", &config_path);
-        cmd.args(["config", "approvals", "clear"])
-            .current_dir(repo.root_path());
-        set_temp_home_env(&mut cmd, temp_home.path());
-
-        assert_cmd_snapshot!(cmd);
-    });
-}
-
-/// Test `wt config approvals clear` when project has approvals to clear.
-#[rstest]
-fn test_hook_clear_with_approvals(repo: TestRepo, temp_home: TempDir) {
-    // Remove origin so project_identifier uses full canonical path
-    repo.run_git(&["remote", "remove", "origin"]);
-
-    // Get the canonical path for the project identifier (escaped for TOML)
-    let project_id_str = repo.project_id();
-
-    // Write approved commands as a sibling config.toml of the test approvals path.
-    // The fallback reads config.toml from the same directory as approvals.toml.
-    let config_path = repo.test_approvals_path().with_file_name("config.toml");
-    fs::write(
-        &config_path,
-        format!(
-            r#"worktree-path = "../{{{{ repo }}}}.{{{{ branch }}}}"
-
-[projects.'{project_id_str}']
-approved-commands = ["cargo build", "cargo test", "npm install"]
-"#
-        ),
-    )
-    .unwrap();
-
-    let settings = setup_snapshot_settings_with_home(&repo, &temp_home);
-    settings.bind(|| {
-        let mut cmd = wt_command();
-        repo.configure_wt_cmd(&mut cmd);
-        cmd.env("WORKTRUNK_CONFIG_PATH", &config_path);
-        cmd.args(["config", "approvals", "clear"])
-            .current_dir(repo.root_path());
-        set_temp_home_env(&mut cmd, temp_home.path());
-
-        assert_cmd_snapshot!(cmd);
-    });
-}
-
 /// Test that syntax errors in templates are shown (not swallowed) with --expanded.
 #[rstest]
 fn test_hook_show_expanded_syntax_error(repo: TestRepo, temp_home: TempDir) {
